requirements.txt
streamlit==1.37.1
pandas==2.2.2
python-docx==1.1.2
bcrypt==4.1.2
SQLAlchemy==2.0.32
